Add interactive chat mode
Chat mode (--chat):
- Multi-turn REPL with linenoise line editing and Tab completion
- Builtin arithmetic evaluator (recursive descent, right-assoc exponent)
- Comparison solver ("4*4 vs 3*3") with flexible prefix stripping
- English number words ("twenty three plus forty two")
- Query rewriting for improved base-model completions
- Turn-boundary detection (TOKEN_ANSWER, question-word tokens)
- Paragraph-based output limiting instead of consecutive newline counting
- SSM state persistence (--state), custom templates (--template)
- REPL commands: /reset, /save, /help, /quit
- Configurable stop conditions: --stop-token, --newline-limit,
  --repeat-limit
- NaN/inf rejection in math evaluator via isfinite() checks

Infrastructure:
- Default template "Question: %s\nDetailed answer:" (temp 0.8)
- Raw token ID validation with proper strtol error checking
- mmap: MAP_POPULATE + posix_fadvise(SEQUENTIAL) + MADV_SEQUENTIAL on Linux
- Prefill allocation overflow checks (size_t multiply chain validation)
- Thread pool generation counter widened from uint32 to uint64
- Metal: release previous mmap buffer before replacing

+++ src/threadpool.c
@@ -55,8 +55,8 @@ typedef struct {
     pthread_t handle;
     int tid;
     int node; /* NUMA node this worker is pinned to */
-    _Atomic unsigned int wake_gen;
-    _Atomic unsigned int last_done_gen;
+    _Atomic uint64_t wake_gen;
+    _Atomic uint64_t last_done_gen;
 } __attribute__((aligned(CACHE_LINE))) worker_t;
 
 /*
@@ -94,8 +94,8 @@ typedef struct {
     /* Dispatch metadata (written by main, read by workers after wake) */
     bm_task_fn_t fn;
     void *arg;
-    int n_active;         /* effective thread count for current dispatch */
-    unsigned int cur_gen; /* monotonic generation counter (main-thread only) */
+    int n_active;     /* effective thread count for current dispatch */
+    uint64_t cur_gen; /* monotonic generation counter (main-thread only) */
 } thread_pool_t;
 
 static thread_pool_t g_pool = {.poll_intensity = -1};
@@ -170,7 +170,7 @@ static void pin_to_cpu(pthread_t thread, int cpu)
 static void *worker_entry(void *arg)
 {
     worker_t *w = (worker_t *) arg;
-    unsigned int my_gen = 0;
+    uint64_t my_gen = 0;
     int worker_idx = w->tid - 1; /* workers[0] has tid=1, etc. */
 
     for (;;) {
@@ -473,7 +473,7 @@ void bm_parallel_for(int n_tasks, bm_task_fn_t fn, void *arg)
      * to the worker that does the matching acquire load on wake_gen.
      * Inactive workers are never signaled -- zero thundering-herd.
      */
-    unsigned int gen = ++g_pool.cur_gen;
+    uint64_t gen = ++g_pool.cur_gen;
     for (int i = 0; i < effective - 1; i++)
         atomic_store_explicit(&g_pool.workers[i].wake_gen, gen,
                               memory_order_release);
@@ -523,7 +523,7 @@ void bm_thread_pool_free(void)
 
     atomic_store_explicit(&g_pool.shutdown, 1, memory_order_relaxed);
     /* Wake ALL workers so they can see the shutdown flag */
-    unsigned int gen = ++g_pool.cur_gen;
+    uint64_t gen = ++g_pool.cur_gen;
     for (int i = 0; i < g_pool.n_workers; i++)
         atomic_store_explicit(&g_pool.workers[i].wake_gen, gen,
                               memory_order_release);
